Move characterize to feature.py
